+++ web/html/src/manager/virtualization/pools/create/pools-create.renderer.tsx
@@ -1,0 +1,18 @@
+import * as React from "react";
+import { PoolsCreate } from "./pools-create";
+import SpaRenderer from "core/spa/spa-renderer";
+import { ActionChain } from "components/action-schedule";
+
+type RendererProps = {
+  serverId: string;
+  localTime: string;
+  timezone: string;
+  actionChains: Array<ActionChain>;
+};
+
+export const renderer = (id: string, { serverId, actionChains, timezone, localTime }: RendererProps) => {
+  SpaRenderer.renderNavigationReact(
+    <PoolsCreate serverId={serverId} actionChains={actionChains} timezone={timezone} localTime={localTime} />,
+    document.getElementById(id)
+  );
+};
